+++ src/server/js/index.js
@@ -5,6 +5,7 @@
 import '../../common/js/index.js';
 
 import './check_npms.js';
+import './event-emitter.js';
 import './functions.js';
 import './methods.js';
 import './publish.js';